// types.ts
export interface Student {
  id: string; // Student Number MDS1...
  name: string;
  class: string;
  gender: 'Male' | 'Female' | 'Other';
  dob: string;
  dateOfAdmission: string;
  club: string;
  guardianName: string;
  residentialAddress: string;
  contact: string;
  status: 'Active' | 'Alumni' | 'Transferred' | 'Dropped out';
  photo?: string;
}
// ...
export interface ExamSession {
  id: string;
  name: string; // e.g., "End of Term 1"
  year: string;
  term: string;
}

export interface GradeScale {
  id: string;
  label: string; // A, B, C...
  minMark: number;
  maxMark: number;
  description: string;
}

export interface StudentMark {
  id: string;
  studentId: string;
  sessionId: string;
  subjectId: string; // Referenced by ID
  score: number;
}

export interface Teacher {
  id: string;
  name: string;
  subject?: string; // Primary subject
  email: string;
  contact: string;
  status: 'Active' | 'On Leave';
}

export interface Subject {
  id: string;
  name: string;
  code: string;
  departmentId?: string;
}

export interface Department {
  id: string;
  name: string;
  hodId: string; // Teacher ID
  teacherIds: string[]; // Teacher IDs
  subjectIds: string[]; // Subject IDs
}

export interface ClassInfo {
  id: string;
  name: string; // e.g. "Grade 10 A"
  gradeTeacherId: string; // Teacher ID
  subjectIds: string[]; // List of subjects taught in this class
}

export interface ClassAllocation {
  id: string;
  classId: string;
  teacherId: string;
  subjectId: string;
}
// ...
export enum ModuleType {
  HOME = 'HOME',
  STUDENTS = 'STUDENTS',
  TEACHERS = 'TEACHERS',
  EXAMS = 'EXAMS',
  FEES = 'FEES',
  ENROLLMENT = 'ENROLLMENT',
  PAYMENT = 'PAYMENT', 
  ADMISSION_RECEIPT = 'ADMISSION_RECEIPT',
  ATTENDANCE = 'ATTENDANCE',
  EMIS = 'EMIS',
  TRANSFER_FORM = 'TRANSFER_FORM',
  REGISTER = 'REGISTER',
  EXAM_ANALYSIS = 'EXAM_ANALYSIS',
  CLASS = 'CLASS',
  SETTINGS = 'SETTINGS',
  RESOURCES = 'RESOURCES',
  ANNOUNCEMENTS = 'ANNOUNCEMENTS',
  PERIODMETER = 'PERIODMETER'
}